Removed jsonquery since we're not using it.
Jsonquery was a namespace I used to simplify json, handling all the
validation of each step in the methods and providing shortcuts to
some more complicated querries.

The validations were not valid for this project as I wanted no logging
or throwing from this area, and also the complex querries were not
necessary.