Filled out most of the client object process.
Noticed that when we get the value back I hit a hard wall of "what
do I do here?" For this program we want to output, but what if
we want to re-use this code in a library? Getting the value is an
asynchronous process, and we don't really want output.

Might have to re-write a similar class for libraries instead of
over-complicating this one with more responsibilities.

Possibly use a signal to announce the new value is available? Or
accept a callback?