Added an error callback to the LocalSocketClient class
We use this callback to override the default error behavior.

The default behavior is to write the error to stderr and carry on.
This program overrides that behavior in the case of both clients we
use. We still output the error to stderr, but then we exit with a
fail status.